requirements.txt
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
Faker==15.0.0
kafka-python==2.0.2
mypy-extensions==0.4.3
orjson==3.8.1
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.3   
python-dateutil==2.8.2
redis==4.3.4
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
wrapt==1.14.1